+++ submit_cgap/submission.py
@@ -293,7 +293,7 @@ def _post_submission(server, keypair, bundle_filename, creation_post_data, submi
     return response
 
 
-def submit_metadata_bundle(bundle_filename, institution, project, server, env, validate_only):
+def submit_metadata_bundle(bundle_filename, institution, project, server, env, validate_only, upload_folder=None):
     """
     Does the core action of submitting a metadata bundle.
 
@@ -303,6 +303,7 @@ def submit_metadata_bundle(bundle_filename, institution, project, server, env, v
     :param server: the server to upload to
     :param env: the beanstalk environment to upload to
     :param validate_only: whether to do stop after validation instead of proceeding to post metadata
+    :param upload_folder: folder in which to find files to upload (default: same as bundle_filename)
     """
 
     with script_catch_errors():
@@ -419,7 +420,7 @@ def submit_metadata_bundle(bundle_filename, institution, project, server, env, v
 
         if outcome == 'success':
             show_section(res, 'upload_info')
-            do_any_uploads(res, keydict=keydict, bundle_filename=bundle_filename)
+            do_any_uploads(res, keydict=keydict, bundle_filename=bundle_filename, upload_folder=upload_folder)
 
         exit(0)
 
@@ -449,17 +450,17 @@ def show_upload_info(uuid, server=None, env=None, keydict=None):
             show("No uploads.")
 
 
-def do_any_uploads(res, keydict, bundle_folder=None, bundle_filename=None):
+def do_any_uploads(res, keydict, upload_folder=None, bundle_filename=None):
     upload_info = get_section(res, 'upload_info')
     if upload_info:
         if yes_or_no("Upload %s?" % n_of(len(upload_info), "file")):
             do_uploads(upload_info, auth=keydict,
-                       folder=bundle_folder or (os.path.dirname(bundle_filename) if bundle_filename else None))
+                       folder=upload_folder or (os.path.dirname(bundle_filename) if bundle_filename else None))
         else:
             show("No uploads attempted.")
 
 
-def resume_uploads(uuid, server=None, env=None, bundle_filename=None, keydict=None):
+def resume_uploads(uuid, server=None, env=None, bundle_filename=None, keydict=None, upload_folder=None):
     """
     Uploads the files associated with a given metadata_bundle. This is useful if you answered "no" to the query
     about uploading your data and then later are ready to do that upload.
@@ -469,6 +470,7 @@ def resume_uploads(uuid, server=None, env=None, bundle_filename=None, keydict=No
     :param env: the beanstalk environment to upload to
     :param bundle_filename: th metadata_bundle file to be uploaded
     :param keydict: keydict-style auth, a dictionary of 'key', 'secret', and 'server'
+    :param upload_folder: folder in which to find files to upload (default: same as bundle_filename)
     """
 
     with script_catch_errors():
@@ -476,11 +478,13 @@ def resume_uploads(uuid, server=None, env=None, bundle_filename=None, keydict=No
         server = resolve_server(server=server, env=env)
         keydict = keydict or get_keydict_for_server(server)
         url = ingestion_submission_item_url(server, uuid)
-        response = requests.get(url, auth=keydict_to_keypair(keydict))
+        keypair = keydict_to_keypair(keydict)
+        response = requests.get(url, auth=keypair)
         response.raise_for_status()
         do_any_uploads(response.json(),
                        keydict=keydict,
-                       bundle_filename=bundle_filename or os.path.abspath(os.path.curdir))
+                       bundle_filename=bundle_filename,
+                       upload_folder=upload_folder)
 
 
 def execute_prearranged_upload(path, upload_credentials):
@@ -507,7 +511,7 @@ def execute_prearranged_upload(path, upload_credentials):
         show("Going to upload %s to %s." % (source, target))
         subprocess.check_call(['aws', 's3', 'cp', '--only-show-errors', source, target], env=env)
     except subprocess.CalledProcessError as e:
-        show("Upload failed with exit code %d" % e.returncode)
+        raise RuntimeError("Upload failed with exit code %d" % e.returncode)
     else:
         end = time.time()
         duration = end - start