+++ submit_cgap/tests/test_submission.py
@@ -4,7 +4,7 @@ import os
 import pytest
 import re
 
-from dcicutils.qa_utils import override_environ, ignored, ControlledTime, MockFileSystem, local_attrs
+from dcicutils.qa_utils import override_environ, ignored, ControlledTime, MockFileSystem, local_attrs, raises_regexp
 from unittest import mock
 from .test_utils import shown_output
 from .. import submission as submission_module
@@ -509,7 +509,7 @@ def test_do_any_uploads():
                 do_any_uploads(
                     res=SOME_UPLOAD_INFO_RESULT,
                     keydict=SOME_KEYDICT,
-                    bundle_folder=SOME_OTHER_BUNDLE_FOLDER,  # rather than bundle_filename
+                    upload_folder=SOME_OTHER_BUNDLE_FOLDER,  # rather than bundle_filename
                 )
                 mock_yes_or_no.assert_called_with("Upload %s files?" % n_uploads)
                 mock_uploads.assert_called_with(
@@ -547,7 +547,8 @@ def test_resume_uploads():
                         mock_do_any_uploads.assert_called_with(
                             some_response_json,
                             keydict=SOME_KEYDICT,
-                            bundle_filename=SOME_BUNDLE_FILENAME
+                            bundle_filename=SOME_BUNDLE_FILENAME,
+                            upload_folder=None,
                         )
 
     with mock.patch.object(submission_module, "script_catch_errors", script_dont_catch_errors):
@@ -599,14 +600,14 @@ def test_execute_prearranged_upload():
         with shown_output() as shown:
             with mock.patch("time.time", MockTime().time):
                 with mock.patch("subprocess.call", return_value=17) as mock_aws_call:
-                    execute_prearranged_upload(path=SOME_FILENAME, upload_credentials=SOME_UPLOAD_CREDENTIALS)
+                    with raises_regexp(RuntimeError, "Upload failed with exit code 17"):
+                        execute_prearranged_upload(path=SOME_FILENAME, upload_credentials=SOME_UPLOAD_CREDENTIALS)
                     mock_aws_call.assert_called_with(
                         ['aws', 's3', 'cp', '--only-show-errors', SOME_FILENAME, SOME_UPLOAD_URL],
                         env=SOME_ENVIRON_WITH_CREDS
                     )
                     assert shown.lines == [
                         "Going to upload some-filename to some-url.",
-                        "Upload failed with exit code 17"
                     ]
 
 
@@ -933,7 +934,8 @@ def test_submit_metadata_bundle_old():
                                                         mock_do_any_uploads.assert_called_with(
                                                             final_res,
                                                             bundle_filename=SOME_BUNDLE_FILENAME,
-                                                            keydict=SOME_KEYDICT
+                                                            keydict=SOME_KEYDICT,
+                                                            upload_folder=None,
                                                         )
         assert shown.lines == [
             'The server http://localhost:7777 recognizes you as J Doe <jdoe@cgap.hms.harvard.edu>.',
@@ -1076,7 +1078,9 @@ def test_submit_metadata_bundle_old():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=False)
+                                                                                   validate_only=False,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except SystemExit as e:  # pragma: no cover
                                                             # This is just in case. In fact it's more likely
                                                             # that a normal 'return' not 'exit' was done.
@@ -1125,7 +1129,9 @@ def test_submit_metadata_bundle_old():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=True)
+                                                                                   validate_only=True,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except SystemExit as e:  # pragma: no cover
                                                             assert e.code == 0
                                                         # It's also OK if it doesn't do an exit(0)
@@ -1174,7 +1180,9 @@ def test_submit_metadata_bundle_old():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=False)
+                                                                                   validate_only=False,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except SystemExit as e:
                                                             # We expect to time out for too many waits before success.
                                                             assert e.code == 1
@@ -1391,7 +1399,9 @@ def test_submit_metadata_bundle_new():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=False)
+                                                                                   validate_only=False,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except SystemExit as e:  # pragma: no cover
                                                             # This is just in case. In fact it's more likely
                                                             # that a normal 'return' not 'exit' was done.
@@ -1401,7 +1411,8 @@ def test_submit_metadata_bundle_new():
                                                         mock_do_any_uploads.assert_called_with(
                                                             final_res,
                                                             bundle_filename=SOME_BUNDLE_FILENAME,
-                                                            keydict=SOME_KEYDICT
+                                                            keydict=SOME_KEYDICT,
+                                                            upload_folder=None,
                                                         )
         assert shown.lines == [
             'The server http://localhost:7777 recognizes you as J Doe <jdoe@cgap.hms.harvard.edu>.',
@@ -1455,7 +1466,9 @@ def test_submit_metadata_bundle_new():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=False)
+                                                                                   validate_only=False,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except Exception as e:
                                                             assert "raised for status" in str(e)
                                                         else:  # pragma: no cover
@@ -1504,7 +1517,9 @@ def test_submit_metadata_bundle_new():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=False)
+                                                                                   validate_only=False,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except Exception as e:
                                                             assert "raised for status" in str(e)
                                                         else:  # pragma: no cover
@@ -1544,7 +1559,9 @@ def test_submit_metadata_bundle_new():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=False)
+                                                                                   validate_only=False,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except SystemExit as e:  # pragma: no cover
                                                             # This is just in case. In fact it's more likely
                                                             # that a normal 'return' not 'exit' was done.
@@ -1593,7 +1610,9 @@ def test_submit_metadata_bundle_new():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=True)
+                                                                                   validate_only=True,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except SystemExit as e:  # pragma: no cover
                                                             assert e.code == 0
                                                         # It's also OK if it doesn't do an exit(0)
@@ -1642,7 +1661,9 @@ def test_submit_metadata_bundle_new():
                                                                                    project=SOME_PROJECT,
                                                                                    server=SOME_SERVER,
                                                                                    env=None,
-                                                                                   validate_only=False)
+                                                                                   validate_only=False,
+                                                                                   upload_folder=None,
+                                                                                   )
                                                         except SystemExit as e:
                                                             # We expect to time out for too many waits before success.
                                                             assert e.code == 1